+++ ak_library_management/models/library_member.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from odoo import models, fields, api
-
+from odoo.exceptions import ValidationError
 
 class LibraryMember(models.Model):
     """
@@ -15,8 +15,9 @@ class LibraryMember(models.Model):
     """
     _name = "library.member"
     _description = "Library Members"
+    _rec_name = "member_id"
 
-    name = fields.Char(string="Member Name", required=True)
+    member_id = fields.Many2one(comodel_name='res.partner', string='Member Name', required=True)
     email = fields.Char(string="Email ID")
     phone = fields.Char(string="Contact Number")
     membership_date = fields.Date(string="Membership Start Date")
@@ -31,3 +32,21 @@ class LibraryMember(models.Model):
             val['membership_no'] = self.env['ir.sequence'].next_by_code('library.member')
         res = super().create(vals_list)
         return res
+
+    def action_send_renewal_mail(self):
+        """
+        this action button method for sending renewl membership mail to the library member.
+        """
+        mail_template = self.env.ref('ak_library_management.email_template_renewal_membership')
+        context = {
+            'default_template_id': mail_template.id
+        }
+        if self.env.user.is_librarian:
+            return {
+                'type': 'ir.actions.act_window',
+                'view_mode': 'form',
+                'res_model': 'mail.compose.message',
+                'target': 'new',
+                'context': context,
+            }
+        raise ValidationError("Only Librarian can send email!!! Contact to the Librarian.")