+++ ak_library_management/models/res_users.py
@@ -10,3 +10,4 @@ class ResUsers(models.Model):
     _inherit = ["res.users"]
 
     is_manager = fields.Boolean(string="Is Manager")
+    is_librarian = fields.Boolean(string="Is Librarian")